Commit before refactoring how the game plays
