+++ lib/cpu.rb
@@ -1,4 +1,7 @@
+require './lib/playability'
+
 class Cpu
+  include Playability
   attr_reader :cpu_board
 
   def initialize(cpu_board)
@@ -6,17 +9,17 @@ class Cpu
     @ships = [Ship.new("cruiser", 3), Ship.new("submarine", 2)]
   end
 
-  def cpu_place_ships
-    @ships.each do |ship_to_place|
-      random_cells = []
-      cpu_cells = @cpu_board.cells.keys
-
-      until @cpu_board.valid_placement?(ship_to_place, random_cells) == true
-        random_cells = cpu_cells.sample(ship_to_place.length)
-      end
-      @cpu_board.place(ship_to_place, random_cells)
-    end
-  end
+  # def cpu_place_ships
+  #   @ships.each do |ship_to_place|
+  #     random_cells = []
+  #     cpu_cells = @cpu_board.cells.keys
+  #
+  #     until @cpu_board.valid_placement?(ship_to_place, random_cells) == true
+  #       random_cells = cpu_cells.sample(ship_to_place.length)
+  #     end
+  #     @cpu_board.place(ship_to_place, random_cells)
+  #   end
+  # end
 
 
 end
